Address review: typed containers in Impl, float distance, remove reinterpret_cast
- VamanaIndexImpl/DynamicVamanaIndexImpl: get_distance and reconstruct_at
  now take typed containers (std::span<const float>, SimpleDataView<float>,
  std::span<const size_t>), matching the existing add() convention. The
  Manager wrappers build the views from raw C-ABI pointers.
- Public API: get_distance returns float* instead of double* to match the
  rest of the runtime API (search distances, etc.).
- Replace unsafe reinterpret_cast<const uint64_t*>(ids) with a type-safe
  std::vector<uint64_t> copy.

+++ bindings/cpp/include/svs/runtime/vamana_index.h
@@ -78,7 +78,7 @@ struct SVS_RUNTIME_API VamanaIndex {
 
     // Compute distance between stored vector `id` and `query` (dim floats).
     virtual Status
-    get_distance(double* distance, size_t id, const float* query) const noexcept = 0;
+    get_distance(float* distance, size_t id, const float* query) const noexcept = 0;
 
     // Reconstruct `n` vectors by ID into `output` buffer (n * dim floats).
     virtual Status reconstruct_at(size_t n, const size_t* ids, float* output) noexcept = 0;
